Moved FilteringCatalog to the MefContrib.Hosting.Filter namespace
